# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Engepe\u00e7as - Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/productReplacements": {
            "get": {
                "tags": [
                    "ProductReplacement"
                ],
                "summary": "Obtenha uma Lista de Produto Substituto.",
                "description": "Obtenha Todas os Produto Substituto",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Quantidade limite de exibi\u00e7\u00e3o",
                        "required": false,
                        "type": "integer",
                        "default": "15"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "P\u00e1gina a ser exibida",
                        "required": false,
                        "type": "integer",
                        "default": "1"
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordena\u00e7\u00e3o do retorno",
                        "required": false,
                        "type": "string",
                        "default": "id"
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Dire\u00e7\u00e3o da ordena\u00e7\u00e3o do retorno",
                        "required": false,
                        "type": "string",
                        "default": "ASC"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "Informe a sele\u00e7\u00e3o de campos que devem retornar da consulta separados por virgula",
                        "required": false,
                        "type": "string",
                        "default": "id, product_id, product_utilization_id, user_hrd_id, code_new, code_old, date_include, code_formatted_old, hrd_D017_Id"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Pesquise por qualquer campo, ao usar este campo as outras consultas ser\u00e3o desconsideradas",
                        "required": false,
                        "type": "string",
                        "default": "Busca"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opera\u00e7\u00e3o realizada com sucesso.",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/ProductReplacement"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "EngepecasAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ProductReplacement"
                ],
                "summary": "Armazene um Produto Substituto Rec\u00e9m-criado no Armazenamento ",
                "description": "Store ProductReplacement",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Produto Substituto que deve ser armazenado ",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/ProductReplacement"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opera\u00e7\u00e3o realizada com sucesso.",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/ProductReplacement"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/productReplacements/{id}": {
            "get": {
                "tags": [
                    "ProductReplacement"
                ],
                "summary": "Exibir o Produto Substituto Especificado.",
                "description": "Obter Produto Substituto.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do Produto Substituto",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opera\u00e7\u00e3o realizada com sucesso.",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/ProductReplacement"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "ProductReplacement"
                ],
                "summary": "Atualize o Produto Substituto Especificado no Armazenamento.",
                "description": "Atualizar Produto Substituto.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do Produto Substituto",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Produto Substituto que deve ser atualizado ",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/ProductReplacement"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opera\u00e7\u00e3o realizada com sucesso",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/ProductReplacement"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "ProductReplacement"
                ],
                "summary": "Remova o Produto Substituto Especificado do Armazenamento.",
                "description": "Exclur Produto Substituto",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do Produto Substituto",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opera\u00e7\u00e3o realizada com sucesso.",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ProductReplacement": {
            "required": [
                "product_id",
                "product_utilization_id",
                "user_hrd_id",
                "hrd_D017_Id"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "product_id": {
                    "description": "D017_D001_Id",
                    "type": "integer",
                    "format": "int32"
                },
                "product_utilization_id": {
                    "description": "D017_C008_Id",
                    "type": "integer",
                    "format": "int32"
                },
                "user_hrd_id": {
                    "description": "D017_C007_Id",
                    "type": "integer",
                    "format": "int32"
                },
                "code_new": {
                    "description": "D017_Codigo_Novo",
                    "type": "string"
                },
                "code_old": {
                    "description": "D017_Codigo_Antigo",
                    "type": "string"
                },
                "date_include": {
                    "description": "D017_Data_Inclusao",
                    "type": "string",
                    "format": "date"
                },
                "code_formatted_old": {
                    "description": "D017_Codigo_Antigo_Formatado",
                    "type": "string"
                },
                "hrd_D017_Id": {
                    "description": "D017_Id",
                    "type": "integer",
                    "format": "int32"
                },
                "created_by": {
                    "description": "created_by",
                    "type": "integer",
                    "format": "int32"
                },
                "updated_by": {
                    "description": "updated_by",
                    "type": "integer",
                    "format": "int32"
                },
                "deleted_by": {
                    "description": "deleted_by",
                    "type": "integer",
                    "format": "int32"
                },
                "deleted_at": {
                    "description": "deleted_at",
                    "type": "string",
                    "format": "date-time"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "EngepecasAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}